backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["ManuelJua <manuelalejandrojuarez@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.12"
google-genai = "^1.18.0"
uvicorn = "^0.34.2"
pypdf = "^5.5.0"
aiohttp = "^3.12.6"
beautifulsoup4 = "^4.13.4"
python-multipart = "^0.0.20"
python-dotenv = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
